requirements.txt
python-telegram-bot==21.4
openai>=1.30.0
fpdf2>=2.7.9
python-dotenv>=1.0.1
pydantic>=2.8.0